+++ codegen-client/src/main/kotlin/software/amazon/smithy/rust/codegen/client/smithy/customizations/MetadataCustomization.kt
@@ -24,7 +24,7 @@ class MetadataCustomization(
     private val runtimeConfig = codegenContext.runtimeConfig
     private val codegenScope by lazy {
         arrayOf(
-            "Metadata" to RuntimeType.operationModule(runtimeConfig).resolve("Metadata"),
+            "Metadata" to RuntimeType.smithyRuntimeApiClient(runtimeConfig).resolve("client::orchestrator::Metadata"),
         )
     }
