+++ examples/Cargo.toml
@@ -1,5 +1,6 @@
 # Without this configuration, the workspace will be read from `rust-runtime`, causing the build to fail.
 [workspace]
+resolver = "2"
 members = [
     "pokemon-service-common",
     "pokemon-service",
