+++ examples/pokemon-service-client-usage/Cargo.toml
@@ -24,6 +24,8 @@ aws-smithy-types = { path = "../../rust-runtime/aws-smithy-types/" }
 # Required for `HyperClientBuilder` in `client-connector` example.
 aws-smithy-runtime = { path = "../../rust-runtime/aws-smithy-runtime/", features=["test-util"] }
 
+# Required for `Metadata` in `custom-header-using-interceptor` example.
+aws-smithy-runtime-api = { path = "../../rust-runtime/aws-smithy-runtime-api/", features=["client"] }
 
 
 hyper = { version = "0.14.25", features = ["client", "full"] }
