+++ examples/pokemon-service-client-usage/examples/response-header-interceptor.rs
@@ -2,6 +2,7 @@
  * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
+use aws_smithy_runtime_api::client::orchestrator::Metadata;
 /// This example demonstrates how response headers can be examined before they are deserialized
 /// into the output type.
 ///
@@ -88,9 +89,7 @@ impl Intercept for ResponseHeaderLoggingInterceptor {
         cfg: &mut ConfigBag,
     ) -> Result<(), BoxError> {
         // `Metadata` in the `ConfigBag` has the operation name in it.
-        let metadata = cfg
-            .load::<aws_smithy_http::operation::Metadata>()
-            .expect("metadata should exist");
+        let metadata = cfg.load::<Metadata>().expect("metadata should exist");
         let operation_name = metadata.name().to_string();
 
         // Get the server side request ID and set it in the RequestID data type
